corsola: add voltorb (rt5682s) ucm2 audio config
in my testing no audio artifacts were showing up on reboot, so we keep
this enabled by default. also, the system runs pulseaudio on debian 13
so lets also include that in the reboot command for the readme. kept
instructions in the readme just in case we do have to disable the config
on boot again.

also, running the commands to enable Speaker and InternalMic via alsaucm
cause my audio to not work until a pulseaudio restart, so lets also put
a notice in the readme for that too.

Signed-off-by: kxtzownsu <kxtzownsu@gmail.com>